Remove deprecated total disk usage soft/hard limit types (#34)
* Remove deprecated total disk usage soft/hard limit types

Why:
Acting on the total usage of the cluster makes no sense now that we are
using KIP-827 to determine the usage of all volumes in the active broker
set. Nodes will enter/exit the active set as a matter of course. So the
aggregate usage will fluctuate. Instead we will limit based on the usage
of each volume going forward. IE throttle if any of the volumes is getting
too full.

See Also:
* https://github.com/strimzi/proposals/blob/main/047-cluster-wide-volume-usage-quota-management.md
* [Add min available bytes storage policy](https://github.com/strimzi/kafka-quotas-plugin/pull/33)

Signed-off-by: Robert Young <robeyoun@redhat.com>

* Remove Strimzi Usage documentation

Why:
It's unknown precisely how Strimzi will integrate with the new version, and
the docs could be confusing as they contradict the warning about it not being
included in Strimzi yet

Signed-off-by: Robert Young <robeyoun@redhat.com>

---------

Signed-off-by: Robert Young <robeyoun@redhat.com>