Parallelise storage check and throttle factor expiry jobs
Why:
We now schedule two jobs with a fixed delay, one to observe the volumes
and update the throttle factor if required and a second to check if the
factor has expired and set a fallback throttle factor if required. This
second job is supposed to be a failsafe in case the volume observation
is taking a long time, delaying the application of stale checking. If
we drive both jobs on one thread then the volume observation job may
block the throttle expiration check job. By running with 2 threads the
expiration check can run.

The factor can now be updated from multiple threads so we need better
control over when it is read and written to. With this change we will
be sure that it has been changed in our thread. Synchronised should
be fine as it's a relatively infrequent operation and we've synchronized
some simple computation.

Signed-off-by: Robert Young <robeyoun@redhat.com>
Co-authored-by: Sam Barker <sam@quadrocket.co.uk>